+++ inst/include/TreeTools/ClusterTable.h
@@ -1,6 +1,7 @@
 #ifndef _TREETOOLS_CLUSTERTABLE_H
 #define _TREETOOLS_CLUSTERTABLE_H
 
+#include <array> /* for array */
 #include <bitset> /* for bitset */
 #include <vector> /* for vector */
 #include <Rcpp/Lightest>
@@ -37,7 +38,48 @@ namespace TreeTools {
 
   constexpr int_fast32_t ct_max_leaves = 16383;
   constexpr int_fast32_t ct_stack_size = 4;
-
+  
+  struct CTEntry {
+    int16 L;
+    int16 R;
+    int16 N;
+    int16 W;
+  };
+  
+  static inline void ct_push(std::array<CTEntry, ct_max_leaves>& S_e,
+                             int16 &Spos,
+                             int16 a, int16 b, int16 c, int16 d) noexcept {
+    ASSERT(Spos < static_cast<int16>(ct_max_leaves));
+    CTEntry &e = S_e[Spos++];
+    e.L = a;
+    e.R = b;
+    e.N = c;
+    e.W = d;
+  }
+  
+  static inline void ct_pop(std::array<CTEntry, ct_max_leaves>& S_e,
+                            int16 &Spos,
+                            int16 &a, int16 &b, int16 &c, int16 &d) noexcept {
+    ASSERT(Spos > 0);
+    CTEntry &e = S_e[--Spos];
+    a = e.L;
+    b = e.R;
+    c = e.N;
+    d = e.W;
+  }
+  
+  static inline void ct_push(CTEntry*& st, int16 a, int16 b, int16 c,
+                             int16 d) noexcept {
+    CTEntry &e = *st++;
+    e.L = a; e.R = b; e.N = c; e.W = d;
+  }
+  
+  static inline void ct_pop(CTEntry*& st, int16 &a, int16 &b, int16 &c, 
+                            int16 &d) noexcept {
+    CTEntry &e = *--st;
+    a = e.L; b = e.R; c = e.N; d = e.W;
+  }
+  
   class ClusterTable {
     
     struct alignas(4) ClusterRow {
@@ -214,18 +256,50 @@ namespace TreeTools {
       return ret;
     }
 
+    
+    // Required by TreeDist 2.9.2
+    // TODO Remove in later version, to prefer is_leaf(int16 v)
     [[nodiscard]] inline bool CLUSTONL(int16* L, int16* R) noexcept {
       return X_left(*L) == *L && X_right(*L) == *R;
     }
-
+    
+    // Required by TreeDist 2.9.2
+    // TODO Remove in later version, to prefer is_leaf(int16 v)
     [[nodiscard]] inline bool CLUSTONR(int16* L, int16* R) noexcept {
       return X_left(*R) == *L && X_right(*R) == *R;
     }
-
+    
+    // Required by TreeDist 2.9.2
+    // TODO Remove in later version, to prefer is_leaf(int16 v)
     [[nodiscard]] inline bool ISCLUST(int16* L, int16* R) noexcept {
       // This function procedure returns value true if cluster <L,R> is in X;
       // otherwise it returns value false
-      return CLUSTONL(L, R) || CLUSTONR(L, R);
+      return CLUSTONL(*L, *R) || CLUSTONR(*L, *R);
+    }
+
+    [[nodiscard]] inline bool CLUSTONL(int16 L, int16 R) noexcept {
+      ASSERT(L > 0 && L <= X_ROWS);
+      const ClusterRow &r = x_rows[L - 1];
+      return r.L == L && r.R == R;
+    }
+    
+    [[nodiscard]] inline bool CLUSTONR(int16 L, int16 R) noexcept {
+      ASSERT(R > 0 && R <= X_ROWS);
+      const ClusterRow &r = x_rows[R - 1];
+      return r.L == L && r.R == R;
+    }
+    
+    [[nodiscard]] inline bool ISCLUST(int16 L, int16 R) noexcept {
+      // This function procedure returns value true if cluster <L,R> is in X;
+      // otherwise it returns value false
+      ASSERT(L > 0 && L <= X_ROWS);
+      const ClusterRow &r_L = x_rows[L - 1];
+      if (r_L.L == L && r_L.R == R) return true;
+      
+      ASSERT(L != R);
+      ASSERT(R > 0 && R <= X_ROWS);
+      const ClusterRow &r_R = x_rows[R - 1];
+      return r_R.L == L && r_R.R == R;
     }
 
     inline void CLEAR() noexcept {
@@ -256,10 +330,10 @@ namespace TreeTools {
     inline void SETSW(int16* L, int16* R) noexcept {
       // If <L,R> is a cluster in X, 
       // this procedure sets the cluster switch for <L,R>.
-      if (CLUSTONL(L, R)) {
+      if (CLUSTONL(*L, *R)) {
         ++n_shared;
         SETSWX(L);
-      } else if (CLUSTONR(L, R)) {
+      } else if (CLUSTONR(*L, *R)) {
         ++n_shared;
         SETSWX(R);
       }